onyx optimizations
SCCS-vsn: 1.3

+++ usr.bin/pascal/libpc/WRITES.c
@@ -9,14 +9,14 @@ WRITES(curfile, d1, d2, d3, d4)
 
 	register struct iorec	*curfile;
 	FILE			*d1;
-	long			d2, d3;
+	int			d2, d3;
 	char			*d4;
 {
 	if (curfile->funit & FREAD) {
 		ERROR(EWRITEIT, curfile->pfname);
 		return;
 	}
-	fwrite(d1, (int)d2, (int)d3, d4);
+	fwrite(d1, d2, d3, d4);
 	if (ferror(curfile->fbuf)) {
 		ERROR(EWRITE, curfile->pfname);
 		return;
